Add patients, see attention order and errors.
Now the user can adds a new patient to the system. When a patient is added, the system adds it to the appropriate queue.
Also, the user can see the order of attention for the unit he wants.
Finally, there was an error in the contains() method of the Queue that was fixed.